==> 2021_FB/6-1_Log-linear_Models.R <==

###############################################################################
## 
## Log-linear models
##
## Creator     : Barum Park
## Last Update : 02/21/2020
##
###############################################################################


# load libraries
library(here)
library(igraph)
library(data.table)
library(ggplot2)
library(doParallel)

if (!exists("n_cores")) 
    stop("cannot find n_cores")


message("\n\n\n***** Fitting log-linear models to mobility data *****\n")

# Read in Data -----------------------------------------------------------------

message("Loading data ...")

# read infomap modules
mod_list = readRDS(here("data", "moving_window", "MV1_NoNEC.rds"))

# read wg scheme (drop _adj in column names)
wg_map = fread(
    here("data", "labmap_NoNEC_with_WG.csv")
) %>%
    setnames(names(.), gsub("_adj", "", names(.)))

# read transition data & assign numeric ids to occupations
dat = fread(
    here("data", "dat_FullSamp_agg_NoNEC.csv")
) %>%
    merge(
        wg_map[, .(occ_label, occ_id)],
        by.x = "occ1",
        by.y = "occ_label",
        all.x = TRUE
    ) %>%
    setnames("occ_id", "occ_num1") %>%
    merge(
        wg_map[, .(occ_label, occ_id)],
        by.x = "occ2",
        by.y = "occ_label",
        all.x = TRUE
    ) %>%
    setnames("occ_id", "occ_num2")

# class names
class_names = c("macro", "meso", "micro")

# keep only occ_label and classes in wg_map
wg_map = wg_map[, .SD, .SDcols = c("occ_label", class_names)]


# Run Log-Linear Models ---------------------------------------------------------

# years to fit
years = dat$int_year %>% unique %>% sort

# create cluster for parallel processing
registerDoParallel(cores = n_cores)

message("\nStart fitting log-linear models for each year ...")
res_mat = foreach(y = years, .combine="rbind") %dopar% {
    
    # extract year
    y_dat = dat[int_year == y, .(occ1, occ2, occ_num1, occ_num2, N)]
    
    # get infomap partition
    mod_memb = data.table(
        occ_label = mod_list[[paste0("midyear_", y)]]$names,
        membership = mod_list[[paste0("midyear_", y)]]$membership
    )
    
    # merge
    y_dat = merge(
        y_dat, 
        mod_memb, 
        by.x = "occ1", 
        by.y = "occ_label",
        all.x = TRUE
    ) %>%
        setnames("membership", "mod1") %>%
        merge( 
            mod_memb, 
            by.x = "occ2", 
            by.y = "occ_label",
            all.x = TRUE
        ) %>%
        setnames("membership", "mod2")
    
    # check missing modules
    na_mods = y_dat[
        is.na(mod1) | is.na(mod2), unique(c(occ1, occ2))
    ]
    
    # check whether missing modules are non-isolates
    if (
        y_dat[occ1 %in% na_mods, any(occ1 != occ2)] ||
        y_dat[occ2 %in% na_mods, any(occ1 != occ2)]
    ) 
        stop("missing values in modules for non-isolated occupations")
    
    # drop isolates
    y_dat = y_dat[!(is.na(mod1) | is.na(mod2))]
    
    # add WG schemes
    y_dat = merge(
        y_dat, 
        wg_map, 
        by.x = "occ1",
        by.y = "occ_label",
        all.x = TRUE
    ) %>%
        setnames(
            class_names, paste0(class_names, 1)
        ) %>%
        merge(
            wg_map, 
            by.x = "occ2",
            by.y = "occ_label",
            all.x = TRUE
        ) %>%
        setnames(
            class_names, paste0(class_names, 2)
        )
    
    # this merge should not lead to any missing values
    if (any(is.na(y_dat)))
        stop("missing values generated by merge with wg_map")
        
    message(
        paste0("Fitting models to data of year ", y, " ...")
    )
    
    # add indicators for diagonals and block-diagonals for partitions
    y_dat[
        , `:=`(
            diag_cells = as.numeric(occ_num1 == occ_num2) * occ_num1,
            same_mod   = as.numeric(mod1 == mod2) * mod1, 
            same_macro = as.numeric(macro1 == macro2) * macro1,
            same_meso  = as.numeric(meso1 == meso2) * meso1,
            same_micro = as.numeric(micro1 == micro2) * micro1
        )
    ]
    
    # create baseline formula
    f = "N ~ factor(diag_cells) + factor(occ_num1) + factor(occ_num2)"
    
    # modules
    fit_mod = glm(
        paste0(f,"+ factor(same_mod)"), 
        data = y_dat,
        family = "poisson"
    )
    
    fit_macro = glm(
        paste0(f,"+ factor(same_macro)"), 
        data = y_dat,
        family = "poisson"
    )
    
    fit_meso = glm(
        paste0(f,"+ factor(same_meso)"), 
        data = y_dat,
        family = "poisson"
    )
    
    fit_micro = glm(
        paste0(f,"+ factor(same_micro)"), 
        data = y_dat,
        family = "poisson"
    )
    
    # create vector of results
    res = c(
        y,
        purrr::map_dbl(
            list(fit_mod, fit_macro, fit_meso, fit_micro), BIC
        ),
        purrr::map_dbl(
            list(fit_mod, fit_macro, fit_meso, fit_micro), AIC
        )
    )
    
    return(res)
    
}

# stop cluster
stopImplicitCluster()

# assign names to results
colnames(res_mat) = c(
    "year",
    paste0("BIC_", c("mod", "macro", "meso", "micro")),
    paste0("AIC_", c("mod", "macro", "meso", "micro"))
)

# generate data
ic_res = data.table(res_mat)

# add scaled BIC and AIC (so that within-year mean is zero)
ic_res[
    , 
    paste0("sBIC_", c("mod", class_names)) := .SD - rowMeans(.SD), 
    .SDcols = paste0("BIC_", c("mod", class_names))
][
    ,
    paste0("sAIC_", c("mod", class_names)) := .SD - rowMeans(.SD), 
    .SDcols = paste0("AIC_", c("mod", class_names))
]

# reshape into long format
ic_df = melt(
    ic_res[
        , 
        c("year", grep("sAIC|sBIC", names(ic_res), value = T)), 
        with = FALSE
    ],
    id.vars = "year",
    variable.name = "stats",
    value.name = "value",
    variable.factor = FALSE
)[
    # which IC?
    , ic := ifelse(grepl("BIC", stats), "BIC", "AIC")
][
    # rename partition column
    , partition := {
        
        x = gsub("^.+_", "", stats)
        ifelse(
            x == "mod", 
            "Mobility",
            paste0(
                toupper(substr(x, 1, 1)),
                substr(x, 2, nchar(x))
            )
        )
    }
]

# change partition column to factor
ic_df[
    , partition := factor(
        partition, 
        levels = c("Mobility", "Macro", "Meso", "Micro"),
        labels = paste0(c("Mobility", "Macro", "Meso", "Micro"), " ")
    )
]
                          

message("\nPlotting (scaled) BIC values for partitions (Figure 7) ...")

pdf(here("output", "figures", "Figure_7.pdf"),
    width = 10, height = 5)
print(
    ggplot(
        ic_df[ic == "BIC"], 
        aes(
            x = year, 
            y = value,
            fill = partition
        )
    ) +
      geom_vline(
          xintercept = seq(1989, 2016) -.5, 
          col = "black",
          lty = 2,
          alpha = .15,
          size = .25
      ) +
      geom_bar(
          stat="identity",
          position = position_dodge(width = .75),
          width=.75,
          col="white",
          alpha=.75
      ) +
      scale_fill_grey(start = 0) +
      scale_x_continuous(
          breaks = 1989:2015,
          limits = c(1988.5, 2015.5),
          expand = c(0, 0)
      ) +
      fnb_theme +
      theme(
          legend.title = element_blank(),
          legend.spacing.x = unit(0.5, "line"),
          legend.position = c(.02, .025),
          legend.justification = c(0, 0),
          legend.direction = "horizontal",
      ) +
      labs(x = "\nYear", y = "BIC (scaled)\n")
)
dev.off()

message(paste0("File Path : ", here("output", "figures", "Figure_7.pdf")))
fwrite(ic_df[ic == "BIC"], here("output", "data", "Figure_7.csv"))
message(paste0("Data Path : ", here("output", "data", "Figure_7.csv")))


message("\nPlotting (scaled) BIC values for partitions (Figure E11) ...")

pdf(here("output", "figures", "Figure_E11.pdf"),
    width=10, height=5)
print(
    ggplot(
        ic_df[ic == "AIC"], 
        aes(
            x = year, 
            y = value,
            fill = partition
        )
    ) +
        geom_vline(
            xintercept = seq(1989, 2016) -.5, 
            col = "black",
            lty = 2,
            alpha = .15,
            size = .25
        ) +
        geom_bar(
            stat="identity",
            position = position_dodge(width = .75),
            width=.75,
            col="white",
            alpha=.75
        ) +
        scale_fill_grey(start = 0) +
        scale_x_continuous(
            breaks = 1989:2015,
            limits = c(1988.5, 2015.5),
            expand = c(0, 0)
        ) +
        fnb_theme +
        theme(
            legend.title = element_blank(),
            legend.spacing.x = unit(0.5, "line"),
            legend.position = c(.02, .025),
            legend.justification = c(0, 0),
            legend.direction = "horizontal",
        ) +
        labs(x = "\nYear", y = "AIC (scaled)\n")
)
dev.off()

message(paste0("File Path : ", here("output", "figures", "Figure_E11.pdf")))
fwrite(ic_df[ic == "AIC"], here("output", "data", "Figure_E11.csv"))
message(paste0("Data Path : ", here("output", "data", "Figure_E11.csv")))


message("\nDone!")
